Update pre check in accordions
