+++ src/applications/check-in/pre-check-in/pages/Confirmation/index.jsx
@@ -6,7 +6,7 @@ import PreCheckinConfirmation from '../../../components/PreCheckinConfirmation';
 
 import { useSendPreCheckInData } from '../../../hooks/useSendPreCheckInData';
 
-import { makeSelectForm, makeSelectVeteranData } from '../../../selectors';
+import { makeSelectVeteranData } from '../../../selectors';
 
 const Confirmation = props => {
   const { router } = props;
@@ -16,14 +16,10 @@ const Confirmation = props => {
   const selectVeteranData = useMemo(makeSelectVeteranData, []);
   const { appointments } = useSelector(selectVeteranData);
 
-  const selectForm = useMemo(makeSelectForm, []);
-  const { data: formData } = useSelector(selectForm);
-
   return (
     <PreCheckinConfirmation
       appointments={appointments}
       isLoading={isLoading}
-      formData={formData}
       router={router}
     />
   );
